+++ src/app/components/modules/Header/Header.tsx
@@ -1,0 +1,26 @@
+"use client"
+
+// CSS Imports
+
+// MUI Imports
+import AppBar from '@mui/material/AppBar'
+import Toolbar from '@mui/material/Toolbar'
+import Typography from '@mui/material/Typography'
+import Button from '@mui/material/Button'
+import IconButton from '@mui/material/IconButton'
+import MenuIcon from '@mui/icons-material/Menu'
+
+export default function Header({ handleOpen }) {
+  return (
+    <AppBar position="static">
+      <Toolbar>
+        <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+          Todo App
+        </Typography>
+        <Button variant="outlined" onClick={handleOpen}> 
+          Add Todo
+        </Button>
+      </Toolbar>
+    </AppBar>
+  )
+}