+++ src/app/layout.tsx
@@ -1,8 +1,14 @@
+
+// CSS imports
 import './globals.css'
+
+// NextJS Imports
 import { Inter } from 'next/font/google'
 
+// Variable Declarations
 const inter = Inter({ subsets: ['latin'] })
 
+// Exports
 export const metadata = {
   title: 'Todo App Local',
   description: 'Todo App practice project built with NextJS, TypeScript and Material UI',
@@ -15,7 +21,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        {children}
+      </body>
     </html>
   )
 }